requirements.txt
openai>=1.0.0
anthropic>=0.25.8
python-dotenv==1.0.1
Flask==3.0.3
PyMySQL>=1.1.0
